DOMCore/tests/submissions/Ms2ger/Element-children.html

<!DOCTYPE html>
<title>HTMLCollection edge cases</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="test"><img><img id=foo><img id=foo></div>
<script>
test(function() {
  var container = document.getElementById("test");
  var result = container.children.item("foo");
  assert_true(result instanceof Element, "Expected an Element.");
  assert_false(result.hasAttribute("id"), "Expected the IDless Element.")
})
</script>
